EventStore event store provider

+++ Source/EventFlow/Configuration/IEventFlowConfiguration.cs
@@ -26,7 +26,7 @@ namespace EventFlow.Configuration
 {
     public interface IEventFlowConfiguration
     {
-        long PopulateReadModelEventPageSize { get; }
+        int PopulateReadModelEventPageSize { get; }
         int NumberOfRetriesOnOptimisticConcurrencyExceptions { get; }
         TimeSpan DelayBeforeRetryOnOptimisticConcurrencyExceptions { get; }
     }
